+++ packages/jaspr/lib/src/ui/styles/properties/unit.dart
@@ -9,6 +9,8 @@ extension UnitExt on num {
 abstract class Unit {
   static const Unit zero = _ZeroUnit();
 
+  static const Unit auto = _AutoUnit();
+
   /// Constructs a [Unit] in the form '100%'
   const factory Unit.percent(double value) = _PercentUnit;
 
@@ -35,7 +37,22 @@ class _ZeroUnit implements Unit {
   String get value => '0';
 
   @override
-  bool operator ==(Object other) => identical(this, other) || other is _Unit && other._value == 0;
+  bool operator ==(Object other) =>
+      identical(this, other) || other is _Unit && other._value == 0;
+
+  @override
+  int get hashCode => 0;
+}
+
+class _AutoUnit implements Unit {
+  const _AutoUnit();
+
+  @override
+  String get value => 'auto';
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) || other is _AutoUnit;
 
   @override
   int get hashCode => 0;
@@ -53,8 +70,12 @@ class _Unit implements Unit {
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      _value == 0 && (other is _ZeroUnit || (other is _Unit && other._value == 0)) ||
-      other is _Unit && runtimeType == other.runtimeType && _unit == other._unit && _value == other._value;
+      _value == 0 &&
+          (other is _ZeroUnit || (other is _Unit && other._value == 0)) ||
+      other is _Unit &&
+          runtimeType == other.runtimeType &&
+          _unit == other._unit &&
+          _value == other._value;
 
   @override
   int get hashCode => _value == 0 ? 0 : _unit.hashCode ^ _value.hashCode;